update: separate fs context into node specific
