+++ packages/auto-drive/package.json
@@ -2,7 +2,6 @@
   "name": "@autonomys/auto-drive",
   "packageManager": "yarn@4.2.2",
   "type": "module",
-  "main": "dist/index.js",
   "version": "1.0.12",
   "author": {
     "name": "Autonomys",
@@ -12,17 +11,35 @@
     "node": ">=20.8.0"
   },
   "scripts": {
-    "build": "tsc"
+    "build": "rollup -c"
+  },
+  "exports": {
+    ".": {
+      "browser": "./dist/browser.js",
+      "node": "./dist/node.js",
+      "types": "./dist/index.d.ts"
+    }
   },
   "devDependencies": {
+    "@prerenderer/rollup-plugin": "^0.3.12",
+    "@rollup/plugin-commonjs": "^28.0.1",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-node-resolve": "^15.3.0",
+    "@rollup/plugin-typescript": "^12.1.1",
     "@types/mime-types": "^2",
+    "rollup": "^4.28.0",
+    "rollup-plugin-jscc": "^2.0.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "tslib": "^2.8.1",
     "typescript": "^5.6.3"
   },
   "dependencies": {
     "@autonomys/auto-dag-data": "^1.0.12",
     "jszip": "^3.10.1",
     "mime-types": "^2.1.35",
+    "process": "^0.11.10",
     "rxjs": "^7.8.1",
+    "stream": "^0.0.3",
     "zod": "^3.23.8"
   },
   "gitHead": "aead187672e1a487b54d0f6b973b6a546e2f7e99"
