+++ packages/auto-drive/src/utils/index.ts
@@ -1,4 +1,2 @@
 export * from './async.js'
-export * from './folder.js'
-export * from './stream.js'
 export * from './types.js'
